Change to double recording and do not run beat tracking for all frames

+++ playback.h
@@ -15,7 +15,7 @@ extern "C" {
 
 void Playback_init(void);
 void Playback_cleanup(void);
-void Playback_playWithDelay(short* frame, int frameSize, long delayTime);
+void Playback_playWithDelay(double* frame, int frameSize, long delayTime);
 
 #ifdef __cplusplus
 }
